+++ templates/publisher/measure.html
@@ -45,10 +45,10 @@
         </div>
       </div>
     </section>
-    {# <div class="row">
+    <div class="row">
       <div class="u-clearfix">
         <h3 class="u-float--left">Daily active devices</h3>
-        <strong class="p-heading--three u-float--right u-no-margin--top">5011</strong>
+        <strong class="p-heading--three u-float--right u-no-margin--top">{{ active_devices_total }}</strong>
       </div>
       <div class="grid u-clearfix">
         <div class="col-3">
@@ -57,8 +57,9 @@
           </select>
         </div>
       </div>
-      <p>GRAPH</p>
+      <div id="active_devices" class="snapcraft-metrics__graph"></div>
     </div>
+    {# 
     <div class="row">
       <div class="grid">
         <div class="col-6">
@@ -93,10 +94,11 @@
 <script src="/static/js/modules/billboard.min.js"></script>
 
 <!-- Illustration of country data -->
-<script src="/static/js/snap-metrics.js"></script>
+<script src="/static/js/dist/snap-metrics.js"></script>
 <script>
-  renderMetrics({
-    'installs': {{ installs_metrics|tojson }}
+  snapcraft.metrics.renderMetrics({
+    'installs': {{ installs_metrics|tojson }},
+    'active_devices': {{ active_devices|tojson }}
   });
 </script>
 {% endblock %}
